pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentshell"
version = "0.1.0"
description = "Secure remote control server for CLI AI agents"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
dependencies = [
    "websockets>=12.0",
    "cryptography>=41.0",
    "argon2-cffi>=23.1",
    "pyotp>=2.9",
    "qrcode>=7.4",
]

[project.scripts]
agentshell = "agentshell.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
